Move AttrHandle and AttrHandleMut to attr module, update more documentation
