+++ src/attr.rs
@@ -9,6 +9,8 @@
 //! [`Nlattr`][crate::genl::Nlattr] types in the
 //! `rtnl.rs` and `genl.rs` modules respectively.
 
+use std::slice::{Iter, IterMut};
+
 use crate::{err::NlError, Buffer, Nl};
 
 /// Trait that defines shared operations for netlink attributes.
@@ -40,5 +42,77 @@ pub trait Attribute<T> {
         R: Nl,
     {
         R::deserialize(self.payload().as_ref()).map_err(NlError::new)
+    }
+}
+
+/// Handle returned for traversing nested attribute structures
+pub enum AttrHandle<'a, O, I> {
+    /// Owned vector
+    Owned(O),
+    /// Vector reference
+    Borrowed(&'a [I]),
+}
+
+impl<'a, O, I> AttrHandle<'a, O, I>
+where
+    O: AsRef<[I]>,
+{
+    /// Create new `AttrHandle`
+    pub fn new(owned: O) -> Self {
+        AttrHandle::Owned(owned)
+    }
+
+    /// Create new borrowed `AttrHandle`
+    pub fn new_borrowed(borrowed: &'a [I]) -> Self {
+        AttrHandle::Borrowed(borrowed)
+    }
+
+    /// Pass back iterator over attributes
+    pub fn iter(&self) -> Iter<I> {
+        self.get_attrs().iter()
+    }
+
+    /// Get the underlying owned value as a reference
+    pub fn get_attrs(&self) -> &[I] {
+        match *self {
+            AttrHandle::Owned(ref o) => o.as_ref(),
+            AttrHandle::Borrowed(b) => b,
+        }
+    }
+}
+
+/// Handle for traversing nested attribute structures mutably
+pub enum AttrHandleMut<'a, O, I> {
+    /// Owned vector
+    Owned(O),
+    /// Vector reference
+    Borrowed(&'a mut [I]),
+}
+
+impl<'a, O, I> AttrHandleMut<'a, O, I>
+where
+    O: AsRef<[I]> + AsMut<[I]>,
+{
+    /// Create new `AttrHandle`
+    pub fn new(owned: O) -> Self {
+        AttrHandleMut::Owned(owned)
+    }
+
+    /// Create new borrowed `AttrHandle`
+    pub fn new_borrowed(borrowed: &'a mut [I]) -> Self {
+        AttrHandleMut::Borrowed(borrowed)
+    }
+
+    /// Pass back iterator over attributes
+    pub fn iter_mut(&mut self) -> IterMut<I> {
+        self.get_mut_attrs().iter_mut()
+    }
+
+    /// Get the underlying owned value as a mutable reference or return `None`
+    pub fn get_mut_attrs(&mut self) -> &mut [I] {
+        match self {
+            AttrHandleMut::Owned(ref mut o) => o.as_mut(),
+            AttrHandleMut::Borrowed(b) => b,
+        }
     }
 }
